+++ Modules/Intent.Modules.Blazor.Components.MudBlazor/Templates/StaticContentTemplateRegistrations/MudBlazorSkillSampleFilesStaticContentTemplateRegistration.cs
@@ -1,13 +1,23 @@
+using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 using Intent.Engine;
+using Intent.Modules.Blazor.Templates.Templates.StaticContentTemplateRegistrations;
+using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.Templates.StaticContent;
+using Intent.Registrations;
 using Intent.RoslynWeaver.Attributes;
+using Intent.Templates;
+using Microsoft.Extensions.FileSystemGlobbing;
+using SearchOption = System.IO.SearchOption;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.Templates.StaticContentTemplateRegistration", Version = "1.0")]
 
 namespace Intent.Modules.Blazor.Components.MudBlazor.Templates.StaticContentTemplateRegistrations
 {
+    [IntentManaged(Mode.Merge, Body = Mode.Merge, Signature = Mode.Fully)]
     public class MudBlazorSkillSampleFilesStaticContentTemplateRegistration : StaticContentTemplateRegistration
     {
         public new const string TemplateId = "Intent.Modules.Blazor.Components.MudBlazor.Templates.StaticContentTemplateRegistrations.MudBlazorSkillSampleFilesStaticContentTemplateRegistration";
@@ -16,7 +26,7 @@ namespace Intent.Modules.Blazor.Components.MudBlazor.Templates.StaticContentTemp
         {
         }
 
-        public override string ContentSubFolder => "SkillSamples";
+        public override string ContentSubFolder => "GeneralSkillSamples";
 
 
         public override string[] BinaryFileGlobbingPatterns => new string[] { "*.jpg", "*.png", "*.xlsx", "*.ico", "*.pdf" };